+++ app/redux/settings/actions.js
@@ -12,12 +12,21 @@ export type UpdateRelayLocationsAction = {
   relayLocations: Array<RelayLocationRedux>,
 };
 
+export type UpdateAutoConnectAction = {
+  type: 'UPDATE_AUTO_CONNECT',
+  autoConnect: boolean,
+};
+
 export type UpdateAllowLanAction = {
   type: 'UPDATE_ALLOW_LAN',
   allowLan: boolean,
 };
 
-export type SettingsAction = UpdateRelayAction | UpdateRelayLocationsAction | UpdateAllowLanAction;
+export type SettingsAction =
+  | UpdateRelayAction
+  | UpdateRelayLocationsAction
+  | UpdateAutoConnectAction
+  | UpdateAllowLanAction;
 
 function updateRelay(relay: RelaySettingsRedux): UpdateRelayAction {
   return {
@@ -35,11 +44,18 @@ function updateRelayLocations(
   };
 }
 
-function updateAllowLan(allowLan: boolean): UpdateAllowLanAction {
+function updateAutoConnect(autoConnect: boolean): UpdateAutoConnectAction {
   return {
-    type: 'UPDATE_ALLOW_LAN',
-    allowLan: allowLan,
+    type: 'UPDATE_AUTO_CONNECT',
+    autoConnect,
   };
 }
 
-export default { updateRelay, updateRelayLocations, updateAllowLan };
+function updateAllowLan(allowLan: boolean): UpdateAllowLanAction {
+  return {
+    type: 'UPDATE_ALLOW_LAN',
+    allowLan,
+  };
+}
+
+export default { updateRelay, updateRelayLocations, updateAutoConnect, updateAllowLan };